pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-collapse"
version = "0.1.0"
description = "Minimal experiment pipeline for quantifying tail-token degradation under synthetic code contamination."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "accelerate>=0.34.0",
  "datasets>=2.21.0",
  "huggingface_hub>=0.24.0",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "packaging>=24.0",
  "pandas>=2.2.0",
  "pyyaml>=6.0.1",
  "scipy>=1.13.0",
  "torch>=2.2.0",
  "tqdm>=4.66.0",
  "transformers>=4.44.0",
]

[project.optional-dependencies]
openai = [
  "openai>=1.40.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

